Initialize i as integer

+++ src/provagit/ProvaGit.java
@@ -15,7 +15,7 @@ public class ProvaGit {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        double i = 2;
+        int i = 2;
         while (true) {
             System.out.println("Ehi ciao: "+i);
             i = i * i;
